requirements.txt
chardet==3.0.4
hiredis==0.2.0
python-dateutil==2.7.2
PyYAML==3.12
tzlocal==1.5.1
cloudinary==1.7.0
celery==4.2.1
pony==0.7.6
pytz==2018.7
redis==3.2.0
requests==2.20.1
socketIO-client-nexus==0.7.6
cx-Oracle==7.0.0
Jinja2==2.10
click==7.0
tqdm==4.31.1
PyMySQL==0.9.3
slackclient==1.3.0
flower
cryptography
-e git+https://intelipowerbuild:ZyC4z3kVRQRR@github.com/Intelipower/python-slack-alert.git#egg=slacknotification
